projects/2025_05_urban_v_rural/pyproject.toml
[project]
name = "2025_05_urban_v_rural"
version = "0.0.1"
description = "Urban vs. rural prices"
classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3.12",
]
requires-python = "==3.12.8"
dependencies = [
  "census>=0.8.24",
  "duckdb>=1.2.2",
  "ipykernel>=6.29.0",
  "polars>=1.27.0",
  "pyarrow>=20.0.0",
  "python-dotenv>=1.1.0",
  "tq @ file:///${PROJECT_ROOT}/../../tq",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=4.0.1",
  "ruff>=0.11.0",
  "sqlfluff>=2.3.5"
]

# Linting and type checking tool configs inherited from the root pyproject.toml
[tool.pyright]
extends = "../../pyproject.toml"

[tool.ruff]
extend = "../../pyproject.toml"

[tool.setuptools.packages.find]
include = ["src*"]
